events: Split MSC4471 module into stream submodules
Follows the ruma-events convention of one module per event type, derived
from the type path: m.stream.subscribe, m.stream.cancel, and
m.stream.update each get their own file under stream/, with the shared
StreamDescriptor at the parent.
